event card refactor
