+++ frontend/components/UI/Drawer.tsx
@@ -1,8 +1,8 @@
 import { KeyboardEvent, MouseEvent, ReactNode, useState } from 'react';
 import { Button, Drawer, IconButton } from '@mui/material';
-import KeyboardDoubleArrowUpIcon from '@mui/icons-material/KeyboardDoubleArrowUp';
+import { KeyboardDoubleArrowUp } from '@mui/icons-material';
 
-export default function DrawerComponent({
+export function DrawerComponent({
   icon,
   buttonText,
   children
@@ -37,7 +37,7 @@ export default function DrawerComponent({
       >
         {children}
         <IconButton onClick={toggleDrawer(false)} aria-label="close drawer" size="large">
-          <KeyboardDoubleArrowUpIcon fontSize="inherit" />
+          <KeyboardDoubleArrowUp fontSize="inherit" />
         </IconButton>
       </Drawer>
     </>
