+++ frontend/components/UI/index.tsx
@@ -2,6 +2,7 @@ export * from '@/components/UI/AutoCompleteMultiple';
 export * from '@/components/UI/AutoCompleteTime';
 export * from '@/components/UI/ButtonGroup';
 export * from '@/components/UI/DataTable';
+export * from '@/components/UI/Drawer';
 export * from '@/components/UI/EventListing';
 export * from '@/components/UI/Modal';
 export * from '@/components/UI/PopOverButton';
