+++ frontend/components/index.tsx
@@ -1,3 +1,4 @@
 export * from '@/components/UI';
 export * from '@/components/ClearFavoritesBtn';
+export * from '@/components/EventCard';
 export * from '@/components/Navigation';
